+++ CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.8.4] - 2020-03-28
+
+### Added
+
+- Added 'test card', run python -m rich
+
+### Changed
+
+- Detected windows terminal, defaulting to colorama support
+
+### Fixed
+
+- Fixed table scaling issue
+
 ## [0.8.3] - 2020-03-27
 
 ### Fixed